Fixed error in compare booleans
